Unwrap awaitable response types in filter invocation
