+++ src/Http/Http.Extensions/test/Microsoft.AspNetCore.Http.Extensions.Tests.csproj
@@ -28,8 +28,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <AssemblyAttribute
-      Include="System.Reflection.AssemblyMetadataAttribute">
+    <AssemblyAttribute Include="System.Reflection.AssemblyMetadataAttribute">
       <_Parameter1>RequestDelegateGeneratorTestBaselines</_Parameter1>
       <_Parameter2>$([System.IO.Path]::Combine($(MSBuildProjectDirectory), "RequestDelegateGenerator", "Baselines"))</_Parameter2>
     </AssemblyAttribute>
